rgw: fix lifecycle transition of encrypted multipart objects
Lifecycle transtion can copy objects to a different storage tier.
When this happens, since the object is repacked, the original
manifest is invalidated.  It is necessary to store a special
"parts_len" attribute to fix this.  There was code in PutObj
to handle this, but that was only used for multisite replication,
it is not used by the lifecycle transisiton code.  This fix
adds similar logic to the lifecycle transition code path to make the
same thing happen.

Fixes: https://tracker.ceph.com/issues/23264

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 77b86722300f2df7c1bf418c2f730b888d77f86a)

Resolves: rhbz#2300284
Resolves: rhbz#2394511