interval_set: Refactor insert to allow overlaps.
The current implementation of interval set only allows inserts to be made when the new interval does not overlap the old.  The "union_insert" method works around this by creating a new interval map, then performing a union of the two sets - an inefficient process.

This commit adapts the insert method to have the flexibility of union_insert, but with the efficiency of the old insert method.

We should add a similar erase method in the future, although there is no need for this PR.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>