osd: EC transaction refactor
This commit refactors the EC write path to perform smaller reads/writes.

Extent Cache
This commit will disable the EC cache, but not remove it. This is intentionally done in a very crude way, where the cache is invalidated after every transaction, using the existing code paths intended for clones.  A future commit will refactor the extent cache such that it can cope with much smaller writes, but this is a much large piece of code.

Switch to shard-address space.
Previously, the EC write path tracked buffers in Rados Object (RO) address space. Much of the read optimisations mean that it makese sense tro track such buffers in the shard space. As such, the code has been refactored to use the new shard_extent_map_t structure added in a previous commit.

Reads
The intention is (see FIXME) that the read path will request the minimal set of buffers required for the read/modify/write. This commit goes as far as passing a suitable data structure from the read path into the RMW path.  Additionally, a minimal read will be performed.

Writes
A minimal set of writes will be performed, but will always be performed to all extents. No code is yet added to allow for the delta parity writes mentioned in the design, although the intention to add such code is considered in places.  This means that writes will NOT be performed to shards where the data has not changed.

Clones
This change means that writes to each shard will no longer cover the same shard address range. Clones are however required to be the same on all shards.  Clones will be submitted to every shard and will cover the same shard address range.  Moreover, since clone complexity does not scale with size, a single clone operation will cover all writes within the shard, even if multiple writes are required.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>