erasure-code: Rewrite the decode_chunks interface for all plugins
Change the decode_chunks interface to take an in map and an out map.
The want_to_read set contains the index of each shard we want to recover.
The in map contains the index and bufferptr to all of the shards
that we have available.
The out map contains the index and bufferptr to an empty buffer for
all of the shards that need to be recovered.
want_to_read may not include every shard from the out map.

Signed-off-by: Jamie Pryde <jamiepry@uk.ibm.com>
Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>