src/osd: Modify scrubbing to correctly verify data when written using the new optimised erasure coding
The new erasure coding has removed the hobj_key from an object which scrubbing previously was checking for. In addition the calculation for size of ondisk_data needed to be modified so scrubbing could verify this, as it is calculated differently and requires the shard it is calculating for with the new erasure coding.

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>