osd: Add nonprimary_shards set to pg_pool_t
EC optimizations pools restrict the choice of primary shard to those
that have a complete log and store object attributes. Shards that cannot
become a primary are listed in the nonprimary_shards set.

For a K+M EC pool the 1st data shard and all M coding parity shards
can be a primary.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>