osd: Force a new interval when allow_ec_optimizations is set
To allow the upgrade of an existing EC pool to start using the
new optimizations we need a way of quiescing I/O and swapping
from the legacy EC code to the new EC code when the pool flag
FLAG_EC_OPTIMIZATIONS is set. We achieve this by making the
setting of this flag trigger a new interval which causes an
on_change notification that stops and re-queues inflight I/O
and starts a full iteration of Peering.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>