osd: object_info_t add shard_versions
EC optimized pools do not always update every shard for write I/Os, this
includes not updating object_info_t (OI attribute). The shard versions
map is used to track the version numbers for shards that are not updated
so it is still possible to work out what needs to be backfilled by
comparing the object_info_t structure store by the primary and the
shard being backfilled.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>