osd: WIP: reworked primary shard selection
This is the start at changes to use pgtemp to select a suitable
primary rather than making changes to OSDMap which would then
require all clients to be running the new code to use the
EC optimized pools.

The concept is that peering can detect whether the currently selected
primary is permitted to be a primary or not, if it is not then it can
use pgtemp to force an alternative priamry to be selected.

The OSDMap code always chooses the first present OSD in pgtemp to become
the primary. EC optimized pools with a K+M profile only permit the 1st
data shard and the M coding parity shards to become the primary. Therefore
we shuffle pgtemp for EC optimized pools placing these shards first and
the other (nonprimary) data shards afterwards.

We apply this shuffling of pgtemp in OSDMonitor when setting pgtemp
and unshuffle it in Peering when setting up the acting set. This
shuffling means we do not have to make any changes to OSDMap and
hence retain backwards compatibility with old clients (who only need
to be able to locate the primary).

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>