osd: EC optimizations overaggresive check for missing objects
Relax an assert in read_log_and_missing for optimized EC
pools. Because the log may not have entries for partial
writes but the missing list is calculated from the full
log the need version for a missing item may be newer than
the lastest log entry for that object.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>