osd: Fix prime_pg_temp to avoid deref null ptr when deleting pool
Regression introduced by optimized EC pools. prime_pg_temp is
calling OSDMap::pgtemp_primaryfirst passing a nullptr for pool
in the scenario where a pool is being deleted which causes
a segmentation fault. If the pool is being deleted we dont
care about the acting set so don't need to call
pgtemp_primaryfirst.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>