Fix bug in 4f601170be54953c8d70ea0df9d42bca5ed7f72e
Bug with routing to an acting set [None,Y,X,X]p(X) for a
3+1 optimzed pool where we want it to choose shard 3
because shard 2 is a non-primary. If we just search the
acting set for the first OSD that matches X we will pick
shard 2, so we have to convert the order to primary's first,
then find the matching OSD and then convert this back to
the normal ordering to get shard 3.

FIXME: Squash with 4f601170be54953c8d70ea0df9d42bca5ed7f72e

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>