osd: EC Optimizations fix peering bug causing unfound objects
Fix some unusual scenarios where peering was incorrectly
declaring that objects were missing:

1. There needs to be partial writes that have been applied but
   not completed. This means that pwlc will not yet have been
   updated.
2. The up set needs to change to a completely disjoint set
   of osds from the old up set. This type of change typically
   happens when a PG splits, or the balancer is changing upmaps.

The new primary shard will get a complete copy of the log
from one of the previous OSDs that wasn't a non-primary, it
will then choose the "best" log which will typically be from
a shard that didn't participate in the partial write and
try and merge these logs together.

proc_master_log will work out that the only difference is the
partial_write and that enough shards applied the partial_write
to allow it to roll-forward. All the acting, recovery and
backfill shards are sent a log that includes this write.

However none of these shards have a copy of the object (they
are all new empty shards) so discover_all_missing has to
check the stray shards to work out where the object will
be recovered from. Because of the partial write
not enough of these claim to have an up to date copy of the
object and its marked as unfound.

The fix makes proc_master_log set up pwlc in exactly the
same way as if the write had been completed when it
rolls forward partial writes and then uses this to update
last_compelte for all the shards that weren't involved in
the partial write.

The fix also means some code in GetMissing which was trying
to do a similar thing for shards that were acting,
recovering or backfilling can be deleted.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>