osd: EC Optimizations proc_master_log bug fixes
There are several problems with how proc_master_log is making
decisions about whether to roll-forward or backwards partial
writes for optimized EC pools.

1. The code was getting confused if the last completed I/O
was a partial write (and had set up pwlc) but this was followed
by one or more full writes that had been applied but not completed
and one or more partial writes. proc_master_log starts at the
first divergence (which will be a partial write) but pwlc will
not yet have been updated to take into account the full
writes.

2. The code was getting confused when a shard was backfilling
and advancing pwlc when the writes are applied (backfilling
shards roll-forward writes immediately) and then in
proc_master_log it starts examining earlier writes and finds
a write that was not fully applied. In this case pwlc is left
ahead of where proc_master_log rewinds the log to.

3. proc_master_log was not only adjusting pwlc but was then
also applying this to adjust last_update and last_complete
for shards. This is too early and results in GetMissing
not getting shards to update their missing list. This was
resulting in different shards reporting different needs for
the same object.

Add a new function condiser_adjusting_pwlc that adjusts
pwlc (either forwards or backwards) before proc_master_log
examines partial writes.

Stop proc_master_log from calling apply_pwlc or update_peer_info,
instead do this in search_for_missing when sending info to
another shard.

Modify how the latest version of pwlc is identified, instead
of using the epoch on toversion use the epoch on fromversion
instead. The toversion is assigned to last_update and last_complete
so this change avoids these being assigned the wrong epoch. In
contrast only the version number of fromversion is used.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>