osdc: Rename SplitRead to SplitIo
Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Fix up upgrade

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

osd: Optimized EC missing list not updated on recovering shard

Shards that are recovering (last_complete != last_update) are using
pwlc to advance last_update for writes that did not effect the shard.
However simply incrementing last_update means that the primary doesnt
send the shard log entries that it missed and consequently it cannot
update its missing list.

If the shard is already missing object X at version V1 and there was
a partial write at V2 that did not update the shard, it does not need
to retain the log entry, but it does need to update the missing list
to say it needs V2 rather than V1. This ensures all shards report
a need for an object at the same version and avoids an assert in
MissingLoc::add_active_missing when the primary is trying to
combine the missing lists from all the shards to work out what has
to be recovered.

The fix is to avoid applying pwlc when last_complete != last_update,
this forces the primary to send the log to the recovering shard
which can then update its missing list (and discarding the log
entries as they are partial writes).

Fixes: https://tracker.ceph.com/issues/73249
Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>

Extra debug

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Add validate step

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Bug fixes

docs: Split Reads design (EC Direct Reads and Replica splits)

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

osd: Check for read instability when doing EC direct reads

can_serve_replica_read(...) previously only applied to replicas as all reads in erasure coding would go through the primary. With the implementation of direct reads in EC, it is no longer the case that reads go through the primary, so we want to expand this logic to ec. The existing logic applies almost perfectly to EC in it's current state, so here we just need to update the function names and debug messages to reflect that this code is now used when going through the direct read path in ec as well.

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>

Split read fix

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Modify IO Sequence Tool so all reads are balanced reads part 2

Jon Bailey

Modify IO Sequence Tool so all reads are balanced reads

Jon Bailey

Fix silly iterator bug

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Debug level to 20

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

FIx performance issues.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

Allow multiple read ops

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

WIP

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

WIP

WIP - replica

WIP - osd direct reads.

This is work in progress and is not final code. However it should work, even with error inject.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>