osdc: SplitOps handle client-provided buffers without replacing them.
Some clients provide buffers to be written to by Objecter through a
(legacy?) field in the op called "outbl" (not to be confused with out_bl).

If those clients provide a buffer of exactly the right size, then the
expectation is that the buffer gets written to without being moved. This change
makes SplitOps mimic the read completion behaviour from Objecter.

The memcpy here is a significant performance impact, however it is thought
that most clients will not use this interface.

There are many comments in this code to explain the reasons, as it is quite
a surprising mechanism.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>